Adding unit tests for events and actions
